--- e2e-tests/spring-web/src/main/java/com/foo/web/examples/spring/base/MainController.java
@@ -1,14 +1,0 @@
-package com.foo.web.examples.spring.base;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-    @GetMapping("/")
-    public String index() {
-        System.out.println("in main controller");
-        return "index.html";
-    }
-}
