+++ e2e-tests/spring-web/src/main/resources/static/dropdown/error.html
