--- e2e-tests/spring-web/src/main/resources/static/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<p>
-    Go To:
-</p>
-<ul>
-    <li><a href="/dropdown"> Dropdown</a></li>
-    <li>Form input(coming soon)</li>
-    <li>Radio buttons (coming soon)</li>
-</ul>
-</body>
-</html>
